SlnHierarchy: Properly handle directory separator.
Each iteration over a path segment added path separator before and after,
yielding doubled path separator and tricking the hierarchy resolution into
believing that all folders up to the drive letter are part of the solution.

The logic in GetRootFolder is fixed, and the test updated so that this case is
properly detected.