Change ProjectConfigurationPlatforms to be postSolution
When running SlnGen on a repository with a mix of .Net Standard and .Net framework the solution file generated would hang on load. After binary searching the changes I found that the solution file generated by Visual Studio has ProjectConfigurationPlatforms as postSolution. Making this one change causes the solution to properly load. This fixes issue #71