Added a --solutiondir command line switch.
Added a --solutiondir command line switch to configure the directory in which the solution file will be generated. The solution file is named based on the name of project in the source directory as before.

The --solutionfile will take precedence over this switch, so if the user provides a value for --solutionfile this will be an absolute path to the sln file and thus we dont use the --solutiondir