Change .NET global tool to be just a shim that runs appropriate version of SlnGen
When evaluating MSBuild projects, its important to load the correct version of MSBuild.  If a user has a specific version of .NET SDK in their global.json, you must load the corresponding version of MSBuild in order to safely evaluate the projects.

This change also uses .NET Framework MSBuild if MSBuild.exe is found on the PATH, otherwise it uses the .NET Core MSBuild if dotnet.exe is on the PATH.

Also added a call to hostfxr in order to give a user a better error message if a .NET SDK cannot be found.