Be more strict about locating MSBuild.exe
If a user has an msbuild.bat on the PATH or some really old version, SlnGen won't work and there is no alternative action a user can do to get unblocked.

This change makes the search for MSBuild.exe more strict by only looking for MSBuild.exe and requiring that it be version 15 or above.

Fixes #278