Do not normalize paths from mapped network drives (#306)
On Windows, the GetFinalPathNameByHandle function returns the UNC path if the specified path is from a mapped network drive.  These paths break Visual Studio so they can't be used.  This change detects if the path was normalized and isn't from a network drive, otherwise the original path is used.

Fixes #305